Create separate file for ``Printer`` base class and ``DotPrinter``, and move ``VCGPrinter`` back into original ``vcgutils.py`` for easier diffing
